Rewrite the testsuite.
The new testsuite uses GTest, and it uses ctags and a gawk script (somebody
should test if it's POSIX awk compatible) that generates the main()
function code from the list of test units to call.

TODO: ctags and gawk should only be needed by the make dist target, not by
users wanting to execute tests.