Avoid using SO_REUSEADDR option.
This was introduced as a workaround for a bug, where disconnected clients
failed to free up the bound ports, and caused a DoS on the long term.

Hopefully, the deep issue has been fixed since and this can be
dropped. Also bump the micro-version number as we're going to rely on the
new behaviour on feng's side.