Improve and simplify sockaddr to hostname parsing.
Rather than have the inet_ntop replacement based on WIN32 definition, check
for the function in configure.ac, then make use of either that or
getnameinfo as needed.

This reduces branching a lot because we now run all the code in a single
Address Family branching, for both sock_ntop_host, inet_ntop and get_port.

Export the new neb_sock_parse_address() function since we'll likely use
that in feng soon.